requirements.txt
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
certifi==2016.8.31
colorama==0.3.7
configparser==3.5.0
decorator==4.0.10
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.1.1
MarkupSafe==0.23
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.2
numpy==1.11.1
pandas==0.18.1
pathlib2==2.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.7
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.6.1
pyzmq==15.4.0
qtconsole==4.2.1
scikits.statsmodels==0.3.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.1
traitlets==4.3.0
wcwidth==0.1.7
wheel==0.24.0
widgetsnbextension==1.2.6
win-unicode-console==0.5